Fix small problem witn DotsGraph tooltip
