+++ README.md
@@ -55,6 +55,7 @@ Una vez esté creada la imagen, se despliega de la siguiente manera:
 * src/api/elastic.js
 * src/api/geoserver.js
 * src/charts/BarStackHorizontal.jsx
+* src/charts/DotsGraph.jsx
 * src/compensator/Drawer.jsx
 * src/compensator/InputCompensation.jsx
 * src/compensator/PopMenu.jsx
